Use improved web backend health check
ELB used to make an HTTP request to check if webserver is health.
If the webserver responded with 200 the ELB considered it healthy.
But it ignored content of response. In some cases the webserver responded with empty content which is definitely
not a healthy state.

Now the healthcheck compares title of the page with expected regexp "TwinDB" which is a value from the database.
If it's present it means that a) webserver works b) database works.