Move keys to JSON
Terraform 12 prints a warning if a variable is not used by a module.
There were secrets in the variables that terraform may reveal.
So, moving the secrets in TF_VAR... to mute the warnings

Terraform 12 don't cast true into an integer anymore. Explicitly specify 1/0 to enable/disable some bits in a repo.